Fixed deployment on /browser
Now the files are deployed with the `browser` prefix.

Use the landing page (catalog) and not the collections endpoint.

Also removed the STAC_VERSION which is not required as the version is
taken from the catalog response.

Added as well the HISTORY_MODE=hash which is specified in the README.md
when deploying on S3 or Apache.